Refactor: remove unnecessary else block
The `else` block has been removed, and the `throw new Exception("Password is incorrect");` statement has been moved outside of the `if` block. This change ensures that the exception is thrown unconditionally if the `if` condition is not met, simplifying the code by eliminating unnecessary braces and the `else` keyword. This refactor makes the code more concise and easier to read.